requirements.txt
Django==5.0.8
mysqlclient==2.2.4
python-dotenv==1.0.1
requests==2.32.3
whitenoise==6.7.0
xlrd==2.0.1

Pillow==10.4.0

